Removed special case tags behaviour
While inserting the <ins> or <del> tags we looked at a list of special tags, and tried to also wrap those in extra <ins> or <del> tags.

This behaviour is inherited from the library that this library is based on, and seems to have some bugs in it.

We got some reports in GitHub about this, and intially I tried to fix the feature.

After realizing the use-case for this is not documented, not in this library or the source library, and the use-case for it is vague, and it being more trouble then it's worth, maybe its best to just completely remove it, reducing the code complexity in the process and also fixing the issues all at the same time.